- Switched to CommonLibSSE-NG as a submodule. - Integrated clib-utils using a vcpkg overlay port. - Updated CMake minimum version to 3.29.0. - Refactored CMakeLists.txt for better flexibility. - Moved .cpp and .cc files to src/, and .hpp and .h files to include/. - Updated PCH.h using the CommonLibSSE-NG sample plugin as a reference.
