+++ CMakePresets.json
@@ -2,7 +2,7 @@
     "version": 2,
     "cmakeMinimumRequired": {
         "major": 3,
-        "minor": 21,
+        "minor": 29,
         "patch": 0
     },
     "configurePresets": [
@@ -21,7 +21,11 @@
                 "VCPKG_TARGET_TRIPLET": "x64-windows-skse",
                 "VCPKG_HOST_TRIPLET": "x64-windows-skse",
                 "VCPKG_OVERLAY_TRIPLETS": "${sourceDir}/cmake",
-                "CMAKE_MSVC_RUNTIME_LIBRARY": "MultiThreaded$<$<CONFIG:Debug>:Debug>DLL"
+                "CMAKE_MSVC_RUNTIME_LIBRARY": "MultiThreaded$<$<CONFIG:Debug>:Debug>DLL",
+                "VCPKG_OVERLAY_PORTS": {
+                    "type": "STRING",
+                    "value": "${sourceDir}/cmake/ports/"
+                }
             }
         },
         {
