--- cmake/headerlist.cmake
@@ -1,6 +1,0 @@
-set(headers ${headers}
-	src/PCH.h 
-    src/log.h
-    src/util.h
-    src/hook.h 
-)
