+++ cmake/ports/clib-util/vcpkg.json
@@ -1,0 +1,7 @@
+{
+  "name": "clib-util",
+  "version-string": "1.4.0",
+  "port-version": 1,
+  "description": "",
+  "homepage": "https://github.com/powerof3/CLibUtil"
+}
