--- cmake/sourcelist.cmake
@@ -1,5 +1,0 @@
-set(sources ${sources}
-    src/plugin.cpp
-    src/hook.cpp
-    src/editorID.hpp
-)
