+++ include/hook.h
@@ -1,7 +1,4 @@
-#include "SKSE/Trampoline.h"
-#include "util.h"
-#include "editorID.hpp"
-#pragma warning(disable: 4100)
+#pragma once
 
 namespace ClassicParalysis
 {
