--- src/editorID.hpp
@@ -1,51 +1,0 @@
-#pragma once
-
-#include <string>
-
-namespace clib_util::editorID
-{
-	using _GetFormEditorID = const char* (*)(std::uint32_t);
-
-	inline std::string get_editorID(const RE::TESForm* a_form)
-	{
-		switch (a_form->GetFormType()) {
-		case RE::FormType::Keyword:
-		case RE::FormType::LocationRefType:
-		case RE::FormType::Action:
-		case RE::FormType::MenuIcon:
-		case RE::FormType::Global:
-		case RE::FormType::HeadPart:
-		case RE::FormType::Race:
-		case RE::FormType::Sound:
-		case RE::FormType::Script:
-		case RE::FormType::Navigation:
-		case RE::FormType::Cell:
-		case RE::FormType::WorldSpace:
-		case RE::FormType::Land:
-		case RE::FormType::NavMesh:
-		case RE::FormType::Dialogue:
-		case RE::FormType::Quest:
-		case RE::FormType::Idle:
-		case RE::FormType::AnimatedObject:
-		case RE::FormType::ImageAdapter:
-		case RE::FormType::VoiceType:
-		case RE::FormType::Ragdoll:
-		case RE::FormType::DefaultObject:
-		case RE::FormType::MusicType:
-		case RE::FormType::StoryManagerBranchNode:
-		case RE::FormType::StoryManagerQuestNode:
-		case RE::FormType::StoryManagerEventNode:
-		case RE::FormType::SoundRecord:
-			return a_form->GetFormEditorID();
-		default:
-			{
-				static auto tweaks = GetModuleHandle(L"po3_Tweaks");
-				static auto func = reinterpret_cast<_GetFormEditorID>(GetProcAddress(tweaks, "GetFormEditorID"));
-				if (func) {
-					return func(a_form->formID);
-				}
-				return {};
-			}
-		}
-	}
-}
