+++ src/hook.cpp
@@ -1,4 +1,5 @@
 #include "hook.h"
+#include "ClibUtil/editorID.hpp"
 
 namespace ClassicParalysis
 {
