+++ vcpkg-configuration.json
@@ -1,18 +1,7 @@
 {
     "default-registry": {
         "kind": "git",
-        "repository": "https://github.com/microsoft/vcpkg.git",
-        "baseline": "cc288af760054fa489574bd8e22d05aa8fa01e5c"
-    },
-    "registries": [
-        {
-            "kind": "git",
-            "repository": "https://gitlab.com/colorglass/vcpkg-colorglass",
-            "baseline": "6309841a1ce770409708a67a9ba5c26c537d2937",
-            "packages": [
-                "commonlibsse-ng",
-                "skse64"
-            ]
-        }
-    ]
+        "repository": "https://github.com/microsoft/vcpkg",
+        "baseline": "2f210a9c13fcf2b0c36b1f2187366feec37e3390"
+    }
 }
